Split object.hpp into multiple header files
